tradeaiapi/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
gunicorn
pillow
